Rename context to createContext
